# Fix performance regressions in #57