--- a/src/types.jl
+++ b/src/types.jl
@@ -20,14 +20,8 @@ Implements the `eltype` method of the *iterator* interface for a subtype of `Fin
 """
 Base.eltype(::Type{T}) where T<:FiniteStateMachineIterator{R} where R = R
 
-function Base.iterate(fsm_iter::FiniteStateMachineIterator)
-  ret = generate(fsm_iter, nothing)
-  ret isa IteratorReturn && return nothing
-  ret
-end
-
-function Base.iterate(fsm_iter::FiniteStateMachineIterator, state)
-  ret = generate(fsm_iter, nothing, state)
-  ret isa IteratorReturn && return nothing
-  ret
+function Base.iterate(fsm_iter::FiniteStateMachineIterator, state=nothing)
+   result = fsm_iter()
+   fsm_iter._state === 0xff && return nothing
+   result, nothing
 end