--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -30,6 +30,7 @@ println("Starting tests with $(Threads.nthreads()) threads out of `Sys.CPU_THREA
 @doset "yieldfrom"
 @doset "typeparams"
 @doset "coverage_preservation"
+@doset "performance"
 VERSION >= v"1.8" && @doset "doctests"
 VERSION >= v"1.8" && @doset "aqua"
 get(ENV,"JET_TEST","")=="true" && @doset "jet"