--- a/test/test_jet.jl
+++ b/test/test_jet.jl
@@ -26,6 +26,6 @@ end
         )
     )
     @show rep
-    @test length(JET.get_reports(rep)) <= 4
+    @test length(JET.get_reports(rep)) <= 5
     @test_broken length(JET.get_reports(rep)) == 0
 end