Quick fix for the performance regression reported in #72: reverts the changes to the definition of the `Base.iterate` overload, no longer delegating to `generate`, which is causing unnecessary allocations.

#### Before (`master`)
```
Direct: 
  31.114 ns (0 allocations: 0 bytes)
ResumableFunctions: 
  5.167 μs (249 allocations: 3.89 KiB)
Channels csize=0: 
  2.256 ms (97 allocations: 2.61 KiB)
Channels csize=1: 
  2.194 ms (19 allocations: 1.44 KiB)
Channels csize=20: 
  177.917 μs (20 allocations: 1.77 KiB)
Channels csize=100: 
  85.167 μs (21 allocations: 3.22 KiB)
Task scheduling
  16.917 μs (86 allocations: 3.38 KiB)
Closure: 
  1.308 μs (82 allocations: 1.28 KiB)
Closure optimised: 
  32.067 ns (0 allocations: 0 bytes)
Closure statemachine: 
  31.742 ns (0 allocations: 0 bytes)
Iteration protocol: 
  32.067 ns (0 allocations: 0 bytes)
```

#### After
```
Direct: 
  32.193 ns (0 allocations: 0 bytes)
ResumableFunctions: 
  32.737 ns (0 allocations: 0 bytes)
Channels csize=0: 
  2.232 ms (97 allocations: 2.61 KiB)
Channels csize=1: 
  2.230 ms (19 allocations: 1.44 KiB)
Channels csize=20: 
  173.458 μs (20 allocations: 1.77 KiB)
Channels csize=100: 
  82.083 μs (21 allocations: 3.22 KiB)
Task scheduling
  17.625 μs (86 allocations: 3.38 KiB)
Closure: 
  1.413 μs (82 allocations: 1.28 KiB)
Closure optimised: 
  32.067 ns (0 allocations: 0 bytes)
Closure statemachine: 
  31.732 ns (0 allocations: 0 bytes)
Iteration protocol: 
  32.067 ns (0 allocations: 0 bytes)
```

Closes #72. Note that the `generate`/`@yieldfrom` paths still cause many allocations, but (a) they're not part of the benchmarks and (b) I haven't been able to come up with a fix for that yet.