@gerlero 2023-09-05
FWIW, I've also tried the code from #67 but this change doesn't seem to affect that at all.