@gerlero 2023-09-05
@Krastanov I understand what JET is complaining about, but I'm not sure what's the right way to make it happy.

What are your thoughts on adding a dummy definition of ``(::FiniteStateMachineIterator)(...)`` that throws?