@Krastanov 2023-09-06
This looks good to me. Let me merge another PR that adds the benchmarks to CI, then I will rerun the CI here and merge.

I will look into the JET report, but unless there is a very clean non-hacky fix, I would just bump the number of disregarded warnings.